requirements.txt
setuptools
python-dotenv
pandas
numpy
pymongo[srv]
pymongo
certifi

# -e .